refactor: Separated the single node error handling and renamed both examples
